app/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
dnspython==2.8.0
dotenv==0.9.9
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.119.0
h11==0.16.0
idna==3.11
passlib==1.7.4
psycopg2-binary==2.9.11
pydantic==2.12.2
pydantic-settings==2.11.0
pydantic_core==2.41.4
pydentic==0.0.1.dev3
PyJWT==2.10.1
python-dotenv==1.1.1
python-stdnum==2.1
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlmodel==0.0.27
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
